requirements.txt
kafka-python==1.4.7
faust==1.10.4
fastapi==0.63.0
fire==0.3.1
uvicorn==0.13.3
sse-starlette==0.6.2
databases==0.4.1
SQLAlchemy==1.3.22
asyncpg==0.21.0
psycopg2-binary==2.8.6
git+https://github.com/ebraminio/aiosseclient
